+++ src/ui/about_view.py
@@ -1,6 +1,7 @@
 from tkinter import *
 import cv2
 
+
 def create_image(path, resize=False):
     image = cv2.imread(path, cv2.IMREAD_ANYCOLOR)
     if resize:
@@ -13,7 +14,8 @@ class AboutView:
 
     def __init__(self):
         self._about_window = None
-        self._calculator_image = create_image('./src/assets/calculator2.png') # TODO kuvan saaminen pythoniin on todella vaikeaa :D
+        # TODO kuvan saaminen pythoniin on todella vaikeaa :D
+        self._calculator_image = create_image('./src/assets/calculator2.png')
 
     def show_about_view(self):
         self._about_window = Toplevel()
@@ -27,7 +29,8 @@ class AboutView:
               font=('Helvetica', 10)).pack()
         Label(self._about_window, text='Author: Olli-Pekka Turtio').pack()
         Label(self._about_window, text='Published: 2022©').pack()
-        Label(self._about_window, image=self._calculator_image).pack(padx=30, pady=35)
+        Label(self._about_window, image=self._calculator_image).pack(
+            padx=30, pady=35)
 
 
 about_view = AboutView()